requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.29
botocore==1.34.29
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
ecdsa==0.18.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
stripe==8.0.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.0
